Refactoring. Added some new validations
