+++ sql_generator/mappings.py
@@ -1,0 +1,5 @@
+constraints = {
+    'pk': 'PRIMARY KEY',
+    'notnull': 'NOT NULL',
+    'unique': 'UNIQUE'
+}
