+++ sql_generator/models.py
@@ -3,6 +3,9 @@ class Entity:
         self.name = name
         self.properties = []
 
+    def add_property(self, prop):
+        self.properties.append(prop)
+
 class Property:
     def __init__(self, name, prop_type):
         self.name = name
