+++ sql_generator/templates/sql_create.template
@@ -17,5 +17,3 @@ CREATE TABLE {{ entity.name }} (
 );
 
 {% endfor %}
-
-# TODO: dodaj zarez posle constraints-a
